Don't prevent calls to mprotect
Don't prevent calls to mprotect while an other thread is entering a
syscall. This happens while testing tor browser based on esr31.

Preventing those calls (or making them wait) is needed to prevent
TOCTTOU attacks from a malicious program, but in our use case (checking
that a program does not unintentionally access network), this should not
be a problem.